Started Updating Old Force Logic
Started to convert the old Force Logic from straight up physics push to an impact position accurate `NaturalMotionMessage` - `ApplyBulletImpulse`. Currently devising a way to get the "strength" of a bullet to allow a dynamic force calculation that isn't broken every time Rockstar adds a new weapon - may not work and may rely on fallbacks and name checks such as "Shotgun" or "Pistol" in the name of a Weapon.

Cleared some old code and added a missing Setting that should've been there far before this version was started.